+++ apl-frontend/electron/main/Electron-Backend/Socket/SocketClient.ts
@@ -1,6 +1,6 @@
 import WebSocket from "ws";
 
-const URL = "wss://dev.chromaserver.net/ws";
+const URL = "wss://apl.chromaserver.net/ws";
 
 export class SocketClient {
   static instance: SocketClient; // Singleton