+++ apl-frontend/electron/main/Electron-Backend/index.ts
@@ -1,3 +1,4 @@
+import { safeStorage, systemPreferences } from "electron";
 import { ankiListeners } from "./AnkiListeners";
 import { dashboardListeners } from "./DashboardListeners";
 import { globalListeners } from "./globalListeners";
@@ -7,6 +8,10 @@ import { settingsListeners } from "./SettingsListeners";
 import { setupListeners } from "./SetupConfigBuilder";
 import { surveyListeners } from "./SurveyListeners";
 import { togglListeners } from "./togglListener";
+import keytar from "keytar";
+import { fetchToken, storeToken } from "./util/auth";
+const SERVICE = "MyElectronApp";
+const ACCOUNT = "my–decryption–key";
 
 export default function registerEvents() {
   globalListeners();
@@ -18,4 +23,6 @@ export default function registerEvents() {
   dashboardListeners();
   reportsListeners();
   settingsListeners();
+
+  (async () => {})();
 }