+++ CHANGELOG.md
@@ -1,2 +1,5 @@
-- Added accent color options
-- Various bug fixes
+**Not updating to this version by June 20th 2025 will break APL**
+
+- Replaced Puppeteer with Playwright
+- Use Webhooks for Discord RPC
+- Added support for adding more than 24 hours in a log